requirements.txt
flask
Flask-WTF
peewee
GitHub-Flask
PyGithub